MOVE utilities library and UUID code to separate .c files
[Marble Editor]
- FIX datatypes for map dimensions

[Marble Engine]
- marble_util_file_open() can now create new files and open existing ones without overwriting them
- ADD "marble_system_cpymem()" function
- ADD "marble_util_init()" function
- Frametime and window update cycle calculation is now done using "marble_util_clock" instead of bare HPC
- REMOVE some global variables
- ADD "marble_util_vec_count()" function
- ADD code for "marble_util_file_read()" function
- ADD "marble_util_file_getinfo()" function

+++ Marble Editor/include/lvlview.hpp
@@ -23,7 +23,7 @@ namespace mbe {
         /*
          *  
          */
-        bool create(int width, int height, QString const &cr_loc, QString const &cr_ident, editoraman *ps_aman) {
+        bool create(uint16_t width, uint16_t height, QString const &cr_loc, QString const &cr_ident, editoraman *ps_aman) {
             if (ps_aman == nullptr)
                 return false;
